Use a request id to identify message belonging to a stream
This was necessary for situations where the server completes a
request and quickly switches to the next from the waiting queue. The
client will send a termination request when it detects that the local
stream ended. That can lead to a race condition where the server
receives the termination request after it switched to the new
stream and then terminates the new stream prematurely.

To avoid that, the server now generates a request id when an HTTP
request comes in and shares that with the client. When the client
needs to send any actions to the server, it includes the request id
and the server will only act if the request id matches, otherwise we
assume tha the message belongs to an already terminated stream
and ignore it.